requirements.txt
tensorflow
Keras
opencv-python-headless
numpy
streamlit
pyngrok
patool
gdown
scikit-learn
sklearn